GS-6 Refactor Tests into folder
